LPS-88725 Moves class to layout-impl
